+++ Source/CesiumRuntime/Private/CesiumMetadataValue.cpp
@@ -1,16 +1,26 @@
-// Copyright 2020-2024 CesiumGS, Inc. and Contributors
+// Copyright 2020-2025 CesiumGS, Inc. and Contributors
 
 #include "CesiumMetadataValue.h"
 #include "CesiumPropertyArray.h"
+#include "CesiumPropertyArrayBlueprintLibrary.h"
 #include "UnrealMetadataConversions.h"
 #include <CesiumGltf/MetadataConversions.h>
 #include <CesiumGltf/PropertyTypeTraits.h>
 
+static FCesiumMetadataValue EmptyMetadataValue = FCesiumMetadataValue();
+
 FCesiumMetadataValue::FCesiumMetadataValue()
-    : _value(swl::monostate{}),
+    : _value(), _arrayValue(), _valueType(), _pEnumDefinition(nullptr) {}
+
+FCesiumMetadataValue::FCesiumMetadataValue(FCesiumPropertyArray&& Array)
+    : _value(),
       _arrayValue(),
       _valueType(),
-      _pEnumDefinition() {}
+      _pEnumDefinition(Array._pEnumDefinition) {
+  FCesiumMetadataValueType elementType = Array._elementType;
+  this->_valueType = {elementType.Type, elementType.ComponentType, true};
+  this->_arrayValue = std::move(Array);
+}
 
 FCesiumMetadataValue::FCesiumMetadataValue(FCesiumMetadataValue&& rhs) =
     default;
@@ -28,6 +38,437 @@ FCesiumMetadataValue&
 FCesiumMetadataValue::operator=(const FCesiumMetadataValue& rhs) {
   *this = FCesiumMetadataValue(rhs);
   return *this;
+}
+
+namespace {
+template <typename TFrom>
+FCesiumMetadataValue convertScalar(
+    const TFrom& input,
+    ECesiumMetadataComponentType targetComponentType) {
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<int8_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Uint8:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<uint8_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Int16:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<int16_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Uint16:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<uint16_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Int32:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<int32_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Uint32:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<uint32_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Int64:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<int64_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Uint64:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<uint64_t, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Float32:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<float, TFrom>::convert(input));
+  case ECesiumMetadataComponentType::Float64:
+    return FCesiumMetadataValue(
+        CesiumGltf::MetadataConversions<double, TFrom>::convert(input));
+  default:
+    return EmptyMetadataValue;
+  }
+}
+
+template <typename T>
+std::vector<T> getScalarArray(const CesiumUtility::JsonValue::Array& array) {
+  std::vector<T> result(array.size());
+
+  for (int64 i = 0; i < int64(result.size()); i++) {
+    std::optional<T> maybeConverted = std::nullopt;
+    if (array[i].isInt64()) {
+      maybeConverted = CesiumGltf::MetadataConversions<T, int64_t>::convert(
+          array[i].getInt64());
+    } else if (array[i].isUint64()) {
+      maybeConverted = CesiumGltf::MetadataConversions<T, uint64_t>::convert(
+          array[i].getUint64());
+    } else if (array[i].isDouble()) {
+      maybeConverted = CesiumGltf::MetadataConversions<T, double>::convert(
+          array[i].getDouble());
+    }
+
+    if (!maybeConverted) {
+      // This value is not a number or could not be converted; exit early.
+      return std::vector<T>();
+    }
+
+    result[i] = *maybeConverted;
+  }
+
+  return result;
+}
+
+FCesiumMetadataValue convertToScalarArray(
+    const CesiumUtility::JsonValue::Array& array,
+    ECesiumMetadataComponentType targetComponentType) {
+  FCesiumPropertyArray resultArray;
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<int8_t>(getScalarArray<int8_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Uint8:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<uint8_t>(getScalarArray<uint8_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Int16:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<int16_t>(getScalarArray<int16_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Uint16:
+    resultArray = FCesiumPropertyArray(CesiumGltf::PropertyArrayCopy<uint16_t>(
+        getScalarArray<uint16_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Int32:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<int32_t>(getScalarArray<int32_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Uint32:
+    resultArray = FCesiumPropertyArray(CesiumGltf::PropertyArrayCopy<uint32_t>(
+        getScalarArray<uint32_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Int64:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<int64_t>(getScalarArray<int64_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Uint64:
+    resultArray = FCesiumPropertyArray(CesiumGltf::PropertyArrayCopy<uint64_t>(
+        getScalarArray<uint64_t>(array)));
+    break;
+  case ECesiumMetadataComponentType::Float32:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<float>(getScalarArray<float>(array)));
+    break;
+  case ECesiumMetadataComponentType::Float64:
+    resultArray = FCesiumPropertyArray(
+        CesiumGltf::PropertyArrayCopy<double>(getScalarArray<double>(array)));
+    break;
+  default:
+    return EmptyMetadataValue;
+  }
+
+  return UCesiumPropertyArrayBlueprintLibrary::GetArraySize(resultArray)
+             ? FCesiumMetadataValue(std::move(resultArray))
+             : EmptyMetadataValue;
+}
+
+template <glm::length_t N, typename T>
+std::optional<glm::vec<N, T>>
+convertToVecN(const CesiumUtility::JsonValue::Array& array) {
+  std::vector<T> components = getScalarArray<T>(array);
+  if (components.size() != N) {
+    return std::nullopt;
+  }
+
+  auto* pResult = reinterpret_cast<glm::vec<N, T>*>(components.data());
+  return *pResult;
+}
+
+template <glm::length_t N>
+FCesiumMetadataValue convertToVecN(
+    const CesiumUtility::JsonValue::Array& array,
+    ECesiumMetadataComponentType targetComponentType) {
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    return FCesiumMetadataValue(convertToVecN<N, int8_t>(array));
+  case ECesiumMetadataComponentType::Uint8:
+    return FCesiumMetadataValue(convertToVecN<N, uint8_t>(array));
+  case ECesiumMetadataComponentType::Int16:
+    return FCesiumMetadataValue(convertToVecN<N, int16_t>(array));
+  case ECesiumMetadataComponentType::Uint16:
+    return FCesiumMetadataValue(convertToVecN<N, uint16_t>(array));
+  case ECesiumMetadataComponentType::Int32:
+    return FCesiumMetadataValue(convertToVecN<N, int32_t>(array));
+  case ECesiumMetadataComponentType::Uint32:
+    return FCesiumMetadataValue(convertToVecN<N, uint32_t>(array));
+  case ECesiumMetadataComponentType::Int64:
+    return FCesiumMetadataValue(convertToVecN<N, int64_t>(array));
+  case ECesiumMetadataComponentType::Uint64:
+    return FCesiumMetadataValue(convertToVecN<N, uint64_t>(array));
+  case ECesiumMetadataComponentType::Float32:
+    return FCesiumMetadataValue(convertToVecN<N, float>(array));
+  case ECesiumMetadataComponentType::Float64:
+    return FCesiumMetadataValue(convertToVecN<N, double>(array));
+  default:
+    return EmptyMetadataValue;
+  }
+}
+
+template <glm::length_t N, typename T>
+std::optional<glm::mat<N, N, T>>
+convertToMatN(const CesiumUtility::JsonValue::Array& array) {
+  std::vector<T> components = getScalarArray<T>(array);
+  if (components.size() != N * N) {
+    return std::nullopt;
+  }
+
+  auto pResult = reinterpret_cast<glm::mat<N, N, T>*>(components.data());
+  return *pResult;
+}
+
+template <glm::length_t N>
+FCesiumMetadataValue convertToMatN(
+    const CesiumUtility::JsonValue::Array& array,
+    ECesiumMetadataComponentType targetComponentType) {
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    return FCesiumMetadataValue(convertToMatN<N, int8_t>(array));
+  case ECesiumMetadataComponentType::Uint8:
+    return FCesiumMetadataValue(convertToMatN<N, uint8_t>(array));
+  case ECesiumMetadataComponentType::Int16:
+    return FCesiumMetadataValue(convertToMatN<N, int16_t>(array));
+  case ECesiumMetadataComponentType::Uint16:
+    return FCesiumMetadataValue(convertToMatN<N, uint16_t>(array));
+  case ECesiumMetadataComponentType::Int32:
+    return FCesiumMetadataValue(convertToMatN<N, int32_t>(array));
+  case ECesiumMetadataComponentType::Uint32:
+    return FCesiumMetadataValue(convertToMatN<N, uint32_t>(array));
+  case ECesiumMetadataComponentType::Int64:
+    return FCesiumMetadataValue(convertToMatN<N, int64_t>(array));
+  case ECesiumMetadataComponentType::Uint64:
+    return FCesiumMetadataValue(convertToMatN<N, uint64_t>(array));
+  case ECesiumMetadataComponentType::Float32:
+    return FCesiumMetadataValue(convertToMatN<N, float>(array));
+  case ECesiumMetadataComponentType::Float64:
+    return FCesiumMetadataValue(convertToMatN<N, double>(array));
+  default:
+    return EmptyMetadataValue;
+  }
+}
+
+template <glm::length_t N, typename T>
+FCesiumMetadataValue
+convertToVecNArray(const CesiumUtility::JsonValue::Array& array) {
+  std::vector<glm::vec<N, T>> values(array.size());
+
+  for (size_t i = 0; i < array.size(); i++) {
+    if (!array[i].isArray() || array[i].getArray().size() != N) {
+      return EmptyMetadataValue;
+    }
+
+    const CesiumUtility::JsonValue::Array& subArray = array[i].getArray();
+    std::optional<glm::vec<N, T>> maybeValue = convertToVecN<N, T>(subArray);
+    if (!maybeValue) {
+      return EmptyMetadataValue;
+    }
+    values[i] = *maybeValue;
+  }
+
+  return FCesiumMetadataValue(
+      CesiumGltf::PropertyArrayCopy<glm::vec<N, T>>(values));
+}
+
+template <glm::length_t N>
+FCesiumMetadataValue convertToVecNArray(
+    const CesiumUtility::JsonValue::Array& array,
+    ECesiumMetadataComponentType targetComponentType) {
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    return FCesiumMetadataValue(convertToVecNArray<N, int8_t>(array));
+  case ECesiumMetadataComponentType::Uint8:
+    return FCesiumMetadataValue(convertToVecNArray<N, uint8_t>(array));
+  case ECesiumMetadataComponentType::Int16:
+    return FCesiumMetadataValue(convertToVecNArray<N, int16_t>(array));
+  case ECesiumMetadataComponentType::Uint16:
+    return FCesiumMetadataValue(convertToVecNArray<N, uint16_t>(array));
+  case ECesiumMetadataComponentType::Int32:
+    return FCesiumMetadataValue(convertToVecNArray<N, int32_t>(array));
+  case ECesiumMetadataComponentType::Uint32:
+    return FCesiumMetadataValue(convertToVecNArray<N, uint32_t>(array));
+  case ECesiumMetadataComponentType::Int64:
+    return FCesiumMetadataValue(convertToVecNArray<N, int64_t>(array));
+  case ECesiumMetadataComponentType::Uint64:
+    return FCesiumMetadataValue(convertToVecNArray<N, uint64_t>(array));
+  case ECesiumMetadataComponentType::Float32:
+    return FCesiumMetadataValue(convertToVecNArray<N, float>(array));
+  case ECesiumMetadataComponentType::Float64:
+    return FCesiumMetadataValue(convertToVecNArray<N, double>(array));
+  default:
+    return EmptyMetadataValue;
+  }
+}
+
+template <glm::length_t N, typename T>
+FCesiumMetadataValue
+convertToMatNArray(const CesiumUtility::JsonValue::Array& array) {
+  std::vector<glm::mat<N, N, T>> values(array.size());
+
+  for (size_t i = 0; i < array.size(); i++) {
+    if (!array[i].isArray() || array[i].getArray().size() != N * N) {
+      return EmptyMetadataValue;
+    }
+
+    const CesiumUtility::JsonValue::Array& subArray = array[i].getArray();
+    std::optional<glm::mat<N, N, T>> maybeValue = convertToMatN<N, T>(subArray);
+    if (!maybeValue) {
+      return EmptyMetadataValue;
+    }
+    values[i] = *maybeValue;
+  }
+
+  return FCesiumMetadataValue(
+      CesiumGltf::PropertyArrayCopy<glm::mat<N, N, T>>(values));
+}
+
+template <glm::length_t N>
+FCesiumMetadataValue convertToMatNArray(
+    const CesiumUtility::JsonValue::Array& array,
+    ECesiumMetadataComponentType targetComponentType) {
+  switch (targetComponentType) {
+  case ECesiumMetadataComponentType::Int8:
+    return FCesiumMetadataValue(convertToMatNArray<N, int8_t>(array));
+  case ECesiumMetadataComponentType::Uint8:
+    return FCesiumMetadataValue(convertToMatNArray<N, uint8_t>(array));
+  case ECesiumMetadataComponentType::Int16:
+    return FCesiumMetadataValue(convertToMatNArray<N, int16_t>(array));
+  case ECesiumMetadataComponentType::Uint16:
+    return FCesiumMetadataValue(convertToMatNArray<N, uint16_t>(array));
+  case ECesiumMetadataComponentType::Int32:
+    return FCesiumMetadataValue(convertToMatNArray<N, int32_t>(array));
+  case ECesiumMetadataComponentType::Uint32:
+    return FCesiumMetadataValue(convertToMatNArray<N, uint32_t>(array));
+  case ECesiumMetadataComponentType::Int64:
+    return FCesiumMetadataValue(convertToMatNArray<N, int64_t>(array));
+  case ECesiumMetadataComponentType::Uint64:
+    return FCesiumMetadataValue(convertToMatNArray<N, uint64_t>(array));
+  case ECesiumMetadataComponentType::Float32:
+    return FCesiumMetadataValue(convertToMatNArray<N, float>(array));
+  case ECesiumMetadataComponentType::Float64:
+    return FCesiumMetadataValue(convertToMatNArray<N, double>(array));
+  default:
+    return EmptyMetadataValue;
+  }
+}
+
+} // namespace
+
+/*static*/ FCesiumMetadataValue FCesiumMetadataValue::fromJsonArray(
+    const CesiumUtility::JsonValue::Array& array,
+    const FCesiumMetadataValueType& targetType) {
+  if (array.empty()) {
+    return EmptyMetadataValue;
+  }
+
+  if (!targetType.bIsArray) {
+    // If the target type is not actually an array, then it must be one of the
+    // following.
+    switch (targetType.Type) {
+    case ECesiumMetadataType::Vec2:
+      return convertToVecN<2>(array, targetType.ComponentType);
+    case ECesiumMetadataType::Vec3:
+      return convertToVecN<3>(array, targetType.ComponentType);
+    case ECesiumMetadataType::Vec4:
+      return convertToVecN<4>(array, targetType.ComponentType);
+    case ECesiumMetadataType::Mat2:
+      return convertToMatN<2>(array, targetType.ComponentType);
+    case ECesiumMetadataType::Mat3:
+      return convertToMatN<3>(array, targetType.ComponentType);
+    case ECesiumMetadataType::Mat4:
+      return convertToMatN<4>(array, targetType.ComponentType);
+      break;
+    default:
+      return EmptyMetadataValue;
+    }
+  }
+
+  if (targetType.Type == ECesiumMetadataType::Boolean) {
+    std::vector<bool> values(array.size());
+    for (size_t i = 0; i < array.size(); i++) {
+      if (!array[i].isBool()) {
+        return EmptyMetadataValue;
+      }
+      values[i] = array[i].getBool();
+    }
+    return FCesiumMetadataValue(CesiumGltf::PropertyArrayCopy<bool>(values));
+  }
+
+  if (targetType.Type == ECesiumMetadataType::String) {
+    std::vector<std::string> values(array.size());
+    for (size_t i = 0; i < array.size(); i++) {
+      if (!array[i].isString()) {
+        return EmptyMetadataValue;
+      }
+      values[i] = array[i].getString();
+    }
+    return FCesiumMetadataValue(
+        CesiumGltf::PropertyArrayCopy<std::string_view>(values));
+  }
+
+  if (targetType.Type == ECesiumMetadataType::Scalar ||
+      targetType.Type == ECesiumMetadataType::Enum) {
+    return convertToScalarArray(array, targetType.ComponentType);
+  }
+
+  // Nested arrays can indicate arrays of vecNs / matNs, and numeric data can be
+  // handled in contiguous buffers.
+  switch (targetType.Type) {
+  case ECesiumMetadataType::Vec2:
+    return convertToVecNArray<2>(array, targetType.ComponentType);
+  case ECesiumMetadataType::Vec3:
+    return convertToVecNArray<3>(array, targetType.ComponentType);
+  case ECesiumMetadataType::Vec4:
+    return convertToVecNArray<4>(array, targetType.ComponentType);
+  case ECesiumMetadataType::Mat2:
+    return convertToMatNArray<2>(array, targetType.ComponentType);
+  case ECesiumMetadataType::Mat3:
+    return convertToMatNArray<3>(array, targetType.ComponentType);
+  case ECesiumMetadataType::Mat4:
+    return convertToMatNArray<4>(array, targetType.ComponentType);
+  default:
+    return EmptyMetadataValue;
+  }
+}
+/*static*/ FCesiumMetadataValue FCesiumMetadataValue::fromJsonValue(
+    const CesiumUtility::JsonValue& jsonValue,
+    const FCesiumMetadataValueType& targetType) {
+  if (jsonValue.isArray()) {
+    return fromJsonArray(jsonValue.getArray(), targetType);
+  } else if (targetType.bIsArray) {
+    return EmptyMetadataValue;
+  }
+
+  if (jsonValue.isBool()) {
+    return targetType.Type == ECesiumMetadataType::Boolean
+               ? FCesiumMetadataValue(jsonValue.getBool())
+               : EmptyMetadataValue;
+  }
+
+  if (jsonValue.isString()) {
+    return targetType.Type == ECesiumMetadataType::String
+               ? FCesiumMetadataValue(std::string_view(jsonValue.getString()))
+               : EmptyMetadataValue;
+  }
+
+  if (targetType.Type == ECesiumMetadataType::Scalar ||
+      targetType.Type == ECesiumMetadataType::Enum) {
+    if (jsonValue.isInt64()) {
+      return FCesiumMetadataValue(
+          convertScalar(jsonValue.getInt64(), targetType.ComponentType));
+    }
+
+    if (jsonValue.isUint64()) {
+      return FCesiumMetadataValue(
+          convertScalar(jsonValue.getUint64(), targetType.ComponentType));
+    }
+
+    if (jsonValue.isDouble()) {
+      return FCesiumMetadataValue(
+          convertScalar(jsonValue.getDouble(), targetType.ComponentType));
+    }
+  }
+
+  return EmptyMetadataValue;
 }
 
 ECesiumMetadataBlueprintType
@@ -263,6 +704,10 @@ FMatrix UCesiumMetadataValueBlueprintLibrary::GetMatrix(
 FString UCesiumMetadataValueBlueprintLibrary::GetString(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FString& DefaultValue) {
+  if (Value._arrayValue) {
+    return UCesiumPropertyArrayBlueprintLibrary::ToString(*Value._arrayValue);
+  }
+
   return swl::visit(
       [&DefaultValue, &Value](auto value) -> FString {
         using ValueType = decltype(value);
@@ -274,7 +719,9 @@ FString UCesiumMetadataValueBlueprintLibrary::GetString(
         } else {
           if constexpr (CesiumGltf::IsMetadataInteger<ValueType>::value) {
             if (Value._pEnumDefinition.IsValid()) {
-              return Value._pEnumDefinition->GetName(value).Get(DefaultValue);
+              TOptional<FString> maybeName =
+                  Value._pEnumDefinition->GetName(value);
+              return maybeName.Get(DefaultValue);
             }
           }